introduce firebase
This commit introduces Firebase authentication and mutli-user support. There
is now a login / signup page upon opening the app.
The app interacts with a non-functional node web server via a retrofit api, but this
code is a temporary placeholder for further firebase storage and will be removed
in future commits.
There is also a different display shown to the user when no art is yet present.
Finally, swipe to delete and swipe to edit functionality has been introduced but not
yet functional.

Signed-off-by: beckyetroy@ <beckyetroy@gmail.com>